Move options event handling to its dedicated file
