+++ src/gui/options_tile.rs
@@ -1,3 +1,11 @@
+use std::sync::{
+	atomic::{AtomicBool, Ordering},
+	mpsc::Sender,
+	Arc,
+};
+
+use crate::enums::Message;
+
 use super::enums::Colors;
 use fltk::{
 	enums::{Color, FrameType},
@@ -31,15 +39,44 @@ pub struct OptionsTile {
 }
 
 impl OptionsTile {
-	pub fn create(x: i32, y: i32) -> Self {
+	pub fn create(x: i32, y: i32, tx: &Sender<Message>, stop_signal: &Arc<AtomicBool>) -> Self {
 		let mut options_tile = Tile::new(x, y, 360, 90, "");
-		let speed_choice = OptionsChoice::create(x + 100, y + 25, 45, 40, "Speed: ", "1|2|3|4");
+		let mut speed_choice = OptionsChoice::create(x + 100, y + 25, 45, 40, "Speed: ", "1|2|3|4");
 
-		let brightness_choice = OptionsChoice::create(x + 100 + 190, y + 25, 45, 40, "Brightness: ", "1|2");
+		let mut brightness_choice = OptionsChoice::create(x + 100 + 190, y + 25, 45, 40, "Brightness: ", "1|2");
 		options_tile.end();
 
 		// Options tile
 		options_tile.set_frame(FrameType::FlatBox);
+
+		speed_choice.set_callback({
+			let tx = tx.clone();
+			let stop_signal = stop_signal.clone();
+			move |choice| {
+				stop_signal.store(true, Ordering::SeqCst);
+				if let Some(value) = choice.choice() {
+					let speed = value.parse::<u8>().unwrap();
+					if (1..=4).contains(&speed) {
+						tx.send(Message::UpdateSpeed { speed }).unwrap();
+					}
+				}
+			}
+		});
+
+		//Brightness
+		brightness_choice.set_callback({
+			let tx = tx.clone();
+			let stop_signal = stop_signal.clone();
+			move |choice| {
+				stop_signal.store(true, Ordering::SeqCst);
+				if let Some(value) = choice.choice() {
+					let brightness = value.parse::<u8>().unwrap();
+					if (1..=2).contains(&brightness) {
+						tx.send(Message::UpdateBrightness { brightness }).unwrap();
+					}
+				}
+			}
+		});
 
 		Self { speed_choice, brightness_choice }
 	}
